added good Folder structure
